requirements.txt
# NeuraFusion Part 3 - Complete Dependencies
# Install with: pip install -r requirements.txt

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers & NLP
transformers>=4.45.0
sentencepiece>=0.1.99
accelerate>=0.24.0

# UI Framework
gradio>=5.0.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Audio Processing
openai-whisper>=20231117
gTTS>=2.5.0
pydub>=0.25.1
soundfile>=0.12.1

# AI Orchestration
langchain>=0.1.0
langchain-community>=0.0.10

# OpenAI Integration (Part 3)
openai>=1.0.0

# Environment Management
python-dotenv>=1.0.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Plotting
plotly>=5.17.0

# Utilities
python-dateutil>=2.8.2
requests>=2.31.0